ESQL: Remove esql-core/test-fixtures
This removes the `esql-core/text-fixtures` project entirely. Most of it
was just used by `esq/qa/testFixtures` so I moved it there. Some of it
was used just by `esql-core/tests`, so I moved it there. The rest was
duplicated by `EsqlTestUtils` so I just deleted it.
