--- x-pack/plugin/esql-core/test-fixtures/src/main/java/org/elasticsearch/xpack/esql/core/TestNode.java
@@ -1,19 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-package org.elasticsearch.xpack.esql.core;
-
-import org.apache.http.HttpHost;
-import org.elasticsearch.TransportVersion;
-import org.elasticsearch.core.Nullable;
-
-public record TestNode(String id, String version, @Nullable TransportVersion transportVersion, HttpHost publishAddress) {
-    @Override
-    public String toString() {
-        return "Node{" + "id='" + id + '\'' + ", version=" + version + '}';
-    }
-}
